// src/app/pages/home/model/totalizador-tempo.ts
export interface TotalizadorTempo {
    tempo_total_dia_veiculos: number;
    tempo_total_hora_veiculos: number;
    tempo_total_minuto_veiculos: number;

    tempo_total_dia_veiculos_movimento: number;
    tempo_total_hora_veiculos_movimento: number;
    tempo_total_minuto_veiculos_movimento: number;

    tempo_total_dia_veiculos_parado: number;
    tempo_total_hora_veiculos_parado: number;
    tempo_total_minuto_veiculos_parado: number;
}